Email: Should back to Email after recording audio as attachment
When attach file in a new email, doAttach() will call
Intent.createChooser() to create an app list for usr
to choose. The Intent being sended to Sound Recorder
doesn't contain value of "exit_after_record". Thus,
when finish recording the audio, Sound Recorder will
not automatically exit and return to Email UI.

Add true value for "exit_after_record" to the Intent
that will be sended to Sound Recorder in doAttach().

CRs-Fixed: 646256

Change-Id: I4b9c846763f5100cdca12cbf7e9df7db22cc8b45