Email: Update the menu after the folder is changed
Since the supported options vary with different mailboxes and folders,
the option menu should changes with the folder. But before the folder
changing is finished, the menu is already prepared. So if the user
tries to click a non-supported menu option item, an exception will be
thrown.

Re-prepare the option menu after the folder is changed.

CRs-fixed: 650798

Change-Id: I931b484e299ed79e38f5651fd57e97f4d7b7253a